Support the case where webpack config is already present next.config.js (#631)
* Support the case where webpack config is already present next.config.js

* Update scripts for experimental-compile/experimental-generate

+++ stack_orchestrator/data/container-build/cerc-nextjs-base/scripts/build-app.sh
@@ -10,21 +10,25 @@ WORK_DIR="${1:-/app}"
 
 cd "${WORK_DIR}" || exit 1
 
-cp next.config.js next.config.dist
+if [ ! -f "next.config.dist" ]; then
+  cp next.config.js next.config.dist
+fi
 
-npm i -g js-beautify
+which js-beautify >/dev/null
+if [ $? -ne 0 ]; then
+  npm i -g js-beautify
+fi
+
 js-beautify next.config.dist > next.config.js
 
-npm install
+WEBPACK_REQ_LINE=$(grep -n "require([\'\"]webpack[\'\"])" next.config.js | cut -d':' -f1)
+if [ -z "$WEBPACK_REQ_LINE" ]; then
+  cat > next.config.js.0 <<EOF
+    const webpack = require('webpack');
+EOF
+fi
 
-CONFIG_LINES=$(wc -l next.config.js | awk '{ print $1 }')
-MOD_EXPORTS_LINE=$(grep -n 'module.exports' next.config.js | cut -d':' -f1)
-
-head -$(( ${MOD_EXPORTS_LINE} - 1 )) next.config.js > next.config.js.1
-
-cat > next.config.js.2 <<EOF
-const webpack = require('webpack');
-
+cat > next.config.js.1 <<EOF
 let envMap;
 try {
   // .env-list.json provides us a list of identifiers which should be replaced at runtime.
@@ -43,21 +47,48 @@ try {
 }
 EOF
 
-grep 'module.exports' next.config.js > next.config.js.3
+CONFIG_LINES=$(wc -l next.config.js | awk '{ print $1 }')
+ENV_LINE=$(grep -n 'env:' next.config.js | cut -d':' -f1)
+WEBPACK_CONF_LINE=$(egrep -n 'webpack:\s+\([^,]+,' next.config.js | cut -d':' -f1)
+NEXT_SECTION_ADJUSTMENT=0
 
-cat > next.config.js.4 <<EOF
-  webpack: (config) => {
-    config.plugins.push(new webpack.DefinePlugin(envMap));
-    return config;
-  },
+if [ -n "$WEBPACK_CONF_LINE" ]; then
+  WEBPACK_CONF_VAR=$(egrep -n 'webpack:\s+\([^,]+,' next.config.js | cut -d',' -f1 | cut -d'(' -f2)
+  head -$(( ${WEBPACK_CONF_LINE} )) next.config.js > next.config.js.2
+  cat > next.config.js.3 <<EOF
+      $WEBPACK_CONF_VAR.plugins.push(new webpack.DefinePlugin(envMap));
 EOF
+  NEXT_SECTION_LINE=$((WEBPACK_CONF_LINE - 1))
+elif [ -n "$ENV_LINE" ]; then
+  head -$(( ${ENV_LINE} - 1 )) next.config.js > next.config.js.2
+  cat > next.config.js.3 <<EOF
+    webpack: (config) => {
+      config.plugins.push(new webpack.DefinePlugin(envMap));
+      return config;
+    },
+EOF
+  NEXT_SECTION_ADJUSTMENT=2
+  NEXT_SECTION_LINE=$ENV_LINE
+else
+  echo "WARNING: Cannot find location to insert environment variable map in next.config.js" 1>&2
+  rm -f next.config.js.*
+  NEXT_SECTION_LINE=0
+fi
 
-tail -$(( ${CONFIG_LINES} - ${MOD_EXPORTS_LINE} + 1 )) next.config.js | grep -v 'process\.env\.' > next.config.js.5
+tail -$(( ${CONFIG_LINES} - ${NEXT_SECTION_LINE} + ${NEXT_SECTION_ADJUSTMENT} )) next.config.js > next.config.js.5
 
-cat next.config.js.* | js-beautify > next.config.js
+cat next.config.js.* | sed 's/^ *//g' | js-beautify | grep -v 'process\.\env\.' | js-beautify > next.config.js
 rm next.config.js.*
 
 "${SCRIPT_DIR}/find-env.sh" "$(pwd)" > .env-list.json
 
-npm run build
-rm .env-list.json
+if [ ! -f "package.dist" ]; then
+  cp package.json package.dist
+fi
+
+cat package.dist | jq '.scripts.cerc_compile = "next experimental-compile"' | jq '.scripts.cerc_generate = "next experimental-generate"' > package.json
+
+npm install || exit 1
+npm run cerc_compile || exit 1
+
+exit 0
